pyproject.toml
[project]
name = "python-assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "python-jose>=3.4.0",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.38",
]
